refactor: update news
