+++ src/actions/news/update.ts
@@ -5,8 +5,21 @@ import { createActionError, createPrismaActionError, createZodActionError } from
 import type { SimpleNewsArticle } from '@/server/news/Types'
 import type { ActionReturn } from '@/actions/Types'
 import type { NewsArticleSchemaType } from './schema'
+import { updateNews } from '@/server/news/update'
 
-export async function publishNews(
+
+export async function updateNewsAction(
+    id: number,
+    rawdata: FormData | NewsArticleSchemaType
+): Promise<ActionReturn<Omit<SimpleNewsArticle, 'coverImage'>>> {
+    //TODO: auth 
+    const parse = newsArticleSchema.safeParse(rawdata)
+    if (!parse.success) return createZodActionError(parse)
+    const data = parse.data
+    return await updateNews(id, data) 
+}
+
+export async function publishNewsAction(
     id: number,
     // disable eslint rule temporarily until todo is resolved
     // eslint-disable-next-line @typescript-eslint/no-unused-vars
@@ -27,40 +40,10 @@ export async function publishNews(
     }
 }
 
-export async function updateNews(
-    id: number,
-    rawdata: FormData | NewsArticleSchemaType
-): Promise<ActionReturn<Omit<SimpleNewsArticle, 'coverImage'>>> {
-    try {
-        const parse = newsArticleSchema.safeParse(rawdata)
-        if (!parse.success) {
-            return createZodActionError(parse)
-        }
-        const data = parse.data
-        const news = await prisma.newsArticle.update({
-            where: { id },
-            data: {
-                description: data.description,
-                article: {
-                    update: {
-                        name: data.name
-                    }
-                },
-                endDateTime: data.endDateTime || undefined,
-            }
-        })
-        return {
-            success: true,
-            data: news
-        }
-    } catch (error) {
-        return createPrismaActionError(error)
-    }
-}
 
 // disable eslint rule temporarily until todo is resolved
 // eslint-disable-next-line @typescript-eslint/no-unused-vars
-export async function updateVisibility(id: number, visible: unknown): Promise<ActionReturn<unknown>> {
+export async function updateVisibilityAction(id: number, visible: unknown): Promise<ActionReturn<unknown>> {
     //TODO: add visible field to news
     return createActionError('UNKNOWN ERROR', 'Not implemented')
 }
